review comments @amassarwi
