+++ app/components/views/TicketsPage/PurchaseTab/PurchaseTab.jsx
@@ -7,7 +7,6 @@ function Purchase() {
   // Legacy hooks - this can be removed after stopping support vsp v1/v2.
   // vsps v3 don't need to import script to configure them.
   const [isShowingVsp, toggleShowVsp] = useState(false);
-  // const [isShowingImportScript, setShowImportScript] = useState(false);
   const { isLegacy, toggleIsLegacy } = usePurchaseTab();
 
   return isLegacy ? (
